gjort om allt mitt till en komponent
